now compiling and running on windows 10 with vs 2017

+++ Game/InputManager.cpp
@@ -121,6 +121,6 @@ void InputManager::clearInput()
 {
 	for (int i = 0; i < 3; i++)
 		mouseButton[i] = false;
-	for (int i = 0; i < keys.size(); i++)
+	for (size_t i = 0; i < keys.size(); i++)
 		keys[i] = false;
 }
